# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI Operational POS API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-UNLICENSED
  description: Phase 0 foundation contract with the backend voice transcription slice. Product, auth, sales, inventory, and intent handlers are not declared until implemented.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
...
  /api/v1/orders:
    get:
      operationId: listOrders
      summary: List past orders
      description: Returns a list of orders optionally filtered by location.
      parameters:
        - name: location_id
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by inventory location UUID.
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createOrder
      summary: Create and complete a POS order
      description: Persists order and items, deducts inventory stock atomically, and appends sale stock movement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderDetail'
        '409':
          description: Insufficient inventory stock for order items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error or insufficient payment amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/customers/{id}:
    get:
      operationId: getCustomerById
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateCustomer
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/pets/{id}:
    get:
      operationId: getPetById
      summary: Get a pet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pet detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '404':
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updatePet
      summary: Update a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPetRequest'
      responses:
        '200':
          description: Pet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '404':
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/services/{id}:
    get:
      operationId: getServiceById
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessService'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateService
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertServiceRequest'
      responses:
        '200':
          description: Service updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessService'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/packages/{id}:
    get:
      operationId: getPackageById
      summary: Get a package by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Package detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPackage'
        '404':
          description: Package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updatePackage
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageRequest'
      responses:
        '200':
          description: Package updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPackage'
        '404':
          description: Package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Payload invalid or service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/bookings/{id}/status:
    post:
      operationId: changeBookingStatus
      summary: Move a booking to proses or batal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingStatusRequest'
      responses:
        '200':
          description: Booking status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Booking already closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Transition invalid (selesai requires payment)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
...
    Order:
      type: object
      required:
        - id
        - order_number
        - location_id
        - status
        - payment_method
        - subtotal_idr
        - tax_idr
        - discount_idr
        - total_idr
        - paid_amount_idr
        - change_amount_idr
        - created_at
      properties:
        id:
          type: string
        order_number:
          type: string
        location_id:
          type: string
        status:
          type: string
          enum: [completed, cancelled, draft]
        payment_method:
          type: string
          enum: [cash, qris, debit_card, credit_card]
        subtotal_idr:
          type: integer
        tax_idr:
          type: integer
        discount_idr:
          type: integer
        total_idr:
          type: integer
        paid_amount_idr:
          type: integer
        change_amount_idr:
          type: integer
        notes:
          type: string
        created_at:
          type: string
          format: date-time
...
    CreateOrderRequest:
      type: object
      required:
        - location_id
        - payment_method
        - paid_amount_idr
        - items
      properties:
        location_id:
          type: string
        payment_method:
          type: string
          enum: [cash, qris, debit_card, credit_card]
        paid_amount_idr:
          type: integer
          minimum: 0
        tax_idr:
          type: [integer, "null"]
        discount_idr:
          type: [integer, "null"]
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItemRequest'
    SuccessOrderList:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        request_id:
          type: string
    SuccessOrderDetail:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/OrderDetail'
        request_id:
          type: string
...
    UpsertCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
...
    SuccessCustomer:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Customer'
        request_id:
          type: string
...
    UpsertPetRequest:
      type: object
      required: [customer_id, name]
      properties:
        customer_id:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
        weight_kg:
          type: number
        color:
          type: string
        allergies:
          type: string
        notes:
          type: string
...
    SuccessPet:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Pet'
        request_id:
          type: string
...
    UpsertServiceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [grooming, klinik, penitipan, lainnya]
        price_idr:
          type: integer
        duration_minutes:
          type: integer
        description:
          type: string
...
    SuccessService:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Service'
        request_id:
          type: string
...
    UpsertPackageRequest:
      type: object
      required: [name, items]
      properties:
        name:
          type: string
        price_idr:
          type: integer
        description:
          type: string
        items:
          type: array
          items:
            type: object
            required: [service_id, sessions_included]
            properties:
              service_id:
                type: string
              sessions_included:
                type: integer
...
    SuccessPackage:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Package'
        request_id:
          type: string
...
    BookingStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [proses, batal]
        notes:
          type: string
...
    SuccessBooking:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Booking'
        request_id:
          type: string
...